+++ code/partida.c
@@ -23,11 +23,11 @@ struct Dibujable* motor_medio = NULL;
 struct Dibujable* motor_fuerte = NULL;
 struct Dibujable* terreno = NULL;
 struct Plataforma* plataformas_partida = NULL;
-uint8_t numero_plataformas = 0;
+uint16_t numero_plataformas = 0;
 
 int combustible = 0;
 uint16_t puntuacion_partida = 0;
-uint8_t fisicas = DESACTIVADAS;
+uint16_t fisicas = DESACTIVADAS;
 
 
 void escalar_escena_partida(float factor_x, float factor_y){
@@ -37,16 +37,16 @@ void escalar_escena_partida(float factor_x, float factor_y){
 		escalar_dibujable_en_escena_dados_ejes(motor_medio, factor_x, factor_y);
 		escalar_dibujable_en_escena_dados_ejes(motor_debil, factor_x, factor_y);
 		escalar_dibujable_en_escena_dados_ejes(nave->objeto, factor_x, factor_y);
-		for(uint8_t i = 0; i < numero_plataformas; i++) {
+		for(uint16_t i = 0; i < numero_plataformas; i++) {
 			escalar_dibujable_en_escena_dados_ejes(plataformas_partida[i].linea, factor_x, factor_y);
-			for(uint8_t j = 0; j < plataformas_partida[i].texto->num_caracteres; j++){
+			for(uint16_t j = 0; j < plataformas_partida[i].texto->num_caracteres; j++){
 				escalar_dibujable_en_escena_dados_ejes(plataformas_partida[i].texto->caracteres[j], factor_x, factor_y);
 			}
 		}
 	}
 }
 
-uint16_t evaluar_aterrizaje(uint8_t bonificador, uint8_t es_arista_aterrizable){
+uint16_t evaluar_aterrizaje(uint16_t bonificador, uint16_t es_arista_aterrizable){
 	uint16_t puntuacion = 0;
 
 	if(es_arista_aterrizable == 1){
@@ -95,9 +95,9 @@ void se_ha_aterrizado(){
 
 void gestionar_colisiones() {
 	struct Arista arista_colision = (struct Arista){0};
-	uint8_t bonificador = 1;
-	uint8_t es_arista_aterrizable = 0;
-	uint8_t se_produce_colision = 0;
+	uint16_t bonificador = 1;
+	uint16_t es_arista_aterrizable = 0;
+	uint16_t se_produce_colision = 0;
 
 	// Comprobar colision con el terreno
 	if(hay_colision(nave->objeto, terreno, &arista_colision)){
@@ -106,7 +106,7 @@ void gestionar_colisiones() {
 		if(es_arista_aterrizable == 1){
 
 			// Si hay colision con el terreno -> evaluar si ha sido colision con plataforma
-			for(uint8_t i = 0; i < numero_plataformas; i++) {
+			for(uint16_t i = 0; i < numero_plataformas; i++) {
 				if(hay_colision(nave->objeto, plataformas_partida[i].linea, &arista_colision)) {
 					// La colision ha sido con una plataforma
 					bonificador = plataformas_partida[i].bonificador;
@@ -129,7 +129,7 @@ void gestionar_colisiones() {
 void dibujar_escena(HDC hdc){
     dibujarDibujable(hdc, nave -> objeto);
 	dibujarDibujable(hdc, terreno);
-	for(uint8_t i = 0; i < numero_plataformas; i++){
+	for(uint16_t i = 0; i < numero_plataformas; i++){
 		dibujar_plataforma(hdc, plataformas_partida[i]);
 	}
 	switch(obtener_propulsor()){
@@ -202,7 +202,7 @@ void dibujarHUD(HDC hdc) {
     destruir_texto(txt);
 }
 
-void rotar_nave(uint8_t direccion){
+void rotar_nave(uint16_t direccion){
 	rotarDibujable(nave -> objeto, direccion);
 	rotarDibujable(motor_debil, direccion);
 	rotarDibujable(motor_medio, direccion);
